+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Reads Google Maps API key from app settings.
+- Fix outbound access host and path for Google Maps API.
 
 ## [0.2.0] - 2020-03-13
 ### Added